Интегрированы изменения из linux версии

+++ Development/Src/Common/MathLib.cpp
@@ -137,12 +137,12 @@ namespace NGTech {
 		return *this;
 	}
 
-	float &Vec2::operator[](ptrdiff_t index) {
+	float &Vec2::operator[](int index) {
 		return *(index + &x);
 	}
 
 
-	float Vec2::operator[](ptrdiff_t index) const {
+	float Vec2::operator[](int index) const {
 		return *(index + &x);
 	}
 
@@ -268,12 +268,12 @@ namespace NGTech {
 		return *this;
 	}
 
-	float& Vec3::operator[](ptrdiff_t index) {
+	float& Vec3::operator[](int index) {
 		return *(index + &x);
 	}
 
 
-	float Vec3::operator[](ptrdiff_t index) const {
+	float Vec3::operator[](int index) const {
 		return *(index + &x);
 	}
 
@@ -429,11 +429,11 @@ namespace NGTech {
 		return *this;
 	}
 
-	float& Vec4::operator[](ptrdiff_t index) {
+	float& Vec4::operator[](int index) {
 		return *(index + &x);
 	}
 
-	float Vec4::operator[](ptrdiff_t index) const {
+	float Vec4::operator[](int index) const {
 		return *(index + &x);
 	}
 
@@ -587,11 +587,11 @@ namespace NGTech {
 		return *this;
 	}
 
-	float &Mat3::operator[](ptrdiff_t index) {
+	float &Mat3::operator[](int index) {
 		return e[index];
 	}
 
-	float Mat3::operator[](ptrdiff_t index) const {
+	float Mat3::operator[](int index) const {
 		return e[index];
 	}
 
@@ -784,11 +784,11 @@ namespace NGTech {
 		return *this;
 	}
 
-	float &Mat4::operator[](ptrdiff_t index) {
+	float &Mat4::operator[](int index) {
 		return e[index];
 	}
 
-	float Mat4::operator[](ptrdiff_t index) const {
+	float Mat4::operator[](int index) const {
 		return e[index];
 	}
 
@@ -1208,11 +1208,11 @@ namespace NGTech {
 		return (float*)&x;
 	}
 
-	float &Quat::operator[](ptrdiff_t i) {
+	float &Quat::operator[](int i) {
 		return ((float*)&x)[i];
 	}
 
-	const float Quat::operator[](ptrdiff_t i) const {
+	const float Quat::operator[](int i) const {
 		return ((float*)&x)[i];
 	}
